+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export { Breadcrumb, BreadcrumbProps } from './Breadcrumb/Breadcrumb';
 export { Link, LinkProps } from './Link/Link';
 export { LinkExternal, LinkExternalProps } from './Link/LinkExternal';
 export {
