+++ src/core/Breadcrumb/Breadcrumb.tsx
@@ -1,39 +1,62 @@
-import React, { Component } from 'react';
+import React, { Component, ReactNode } from 'react';
 import { default as styled } from 'styled-components';
+import classnames from 'classnames';
 import { getConditionalAriaProp } from '../../utils/aria';
-import {
-  Breadcrumb as CompBreadcrumb,
-  BreadcrumbProps as CompBreadcrumbProps,
-} from '../../components/Breadcrumb/Breadcrumb';
 import { LinkProps } from '../Link/Link';
 import { BreadcrumbLink, BreadcrumbLinkProps } from './BreadcrumbLink';
+import { HtmlLi, HtmlNav, HtmlNavProps, HtmlOl } from '../../reset';
 import { baseStyles } from './Breadcrumb.baseStyles';
 
 type BreadcrumbVariant = 'default' | 'link';
+const baseClassName = 'fi-breadcrumb';
+const listClassName = `${baseClassName}_list`;
+const itemClassName = `${baseClassName}_item`;
 
-export interface BreadcrumbProps extends CompBreadcrumbProps {
+export interface BreadcrumbProps extends HtmlNavProps {
+  /** Name for the breadcrumb like 'Breadcrumb' */
+  'aria-label': string;
+  /** Custom classname to extend or customize */
+  className?: string;
   /**
-   * 'default' | 'expansion'
+   * Link element displayed content
+   */
+  children?: ReactNode;
+  /**
+   * 'default' | 'link'
    * @default default
    */
   variant?: BreadcrumbVariant;
 }
 
-const StyledBreadcrumb = styled(CompBreadcrumb)`
-  ${baseStyles};
-`;
-
 type VariantBreadcrumbProps =
   | BreadcrumbProps
   | (LinkProps & { variant?: BreadcrumbVariant });
+
+const breadcrumbItems = (children: ReactNode) =>
+  React.Children.map(children, (child) => (
+    <HtmlLi className={itemClassName}>{child}</HtmlLi>
+  ));
+
+class BaseBreadcrumb extends Component<BreadcrumbProps> {
+  render() {
+    const { className, children, ...passProps } = this.props;
+    return (
+      <HtmlNav {...passProps} className={classnames(baseClassName, className)}>
+        <HtmlOl className={listClassName}>{breadcrumbItems(children)}</HtmlOl>
+      </HtmlNav>
+    );
+  }
+}
+
+const StyledBreadcrumb = styled(BaseBreadcrumb)`
+  ${baseStyles};
+`;
 
 /**
  * <i class="semantics" />
  * Used for navigation path
  */
 export class Breadcrumb extends Component<VariantBreadcrumbProps> {
-  static link = (props: BreadcrumbLinkProps) => <BreadcrumbLink {...props} />;
-
   render() {
     const { variant, 'aria-label': ariaLabel, ...passProps } = this.props;
     if (variant === 'link') {
